Make hack/verify.sh make a temp dir in /tmp/
This is important when we add the verify step for Protobuf. That verify
needs to run in Docker, and so needs to mount the temp directory into a
Docker container. On macOS only some paths are mountable by default,
and /tmp/ is one of them. Previously this command, on macOS, was using a
temp dir under /var/.