Add cron job that does defrag without considering threshold
It can be beneficial to defrag even if the threshold is not yet met, so
we add a command line flag that can take the values hourly, daily or
weekly and will force a defrag at this interval. The default value is
set to weekly.
The implementation is not robust, defrags can be missed due to
restarting etc.

A new failure scenario is added as we validate the interval and fail to
start if the value is invalid.